v0.092
Added GuiText and GuiWindow (subclasses of GuiObject). See LevelEdit
project for example of implementation. This commit marks the merging of
LevelEdit and FirstPyGame repos. They have been merged as LevelEdit
strongly depends on FirstPyGames code and it will be a lot easier to
organize with LevelEdit as a subdir of FirstPyGame. This also allows
both projects to work off of the same data directory. (No more multiple
copies of data files)