v0.92a
Created factory for gui objects. Parameter passing for gui objects now
use a dictionary. Added GuiButton (See LevelEdit project for
implementation) Added gui_objects_at(x,y) to check if gui objects are
clicked on. This will be checked for every active/visible gui object on
every pygame.MOUSEBUTTONUP event. If a gui object is clicked,
pygame.MOUSEBUTTONUP is passed to object's input function. Added
pygame.rect to base GuiObject. The base class will set w/h to 1. It is
up to be subclass to set w/h of whatever space it would like to
represent when it is appropriate to do so (ex-after text is rendered)