v0.93a
Added editing and deleting of level objects (F5, F6). Added selection
offset so when an object is selected it doesn't move to the top left
corner. Added a special keybind for GuiState that toggles the state's
activeness. Added is_focus variable for GuiObject to know when it is
focused on. Textboxes now have a different background color when focused
on.